wxauto/requirements.txt
httpx>=0.27.0
pydantic>=2.7.0
typer>=0.12.3
python-dotenv>=1.0.1
qrcode>=7.4.2
